correctionType

+++ src/BmiService.java
@@ -1,8 +1,8 @@
 public class BmiService {
-    public double calculate(double height, double weight) {
+    public int calculate(double height, int weight) {
         double result;
         result = weight / height / height;
-        return result;
+        return (int) result;
     }
 
 }
